+++ DivisiBill/Models/Meal.Persistence.cs
@@ -1,6 +1,7 @@
 using CommunityToolkit.Mvvm.ComponentModel;
 using DivisiBill.Services;
 using System.Diagnostics;
+using System.IO.Compression;
 using System.Runtime.CompilerServices;
 using System.Text;
 using System.Xml;
@@ -15,6 +16,32 @@ namespace DivisiBill.Models;
 /// </summary>
 public partial class Meal : ObservableObjectPlus
 {
+    public static void InitializeFolders()
+    {
+        Directory.CreateDirectory(MealFolderPath);
+        Directory.CreateDirectory(SuspectFolderPath);
+        Directory.CreateDirectory(DeletedItemFolderPath);
+        Directory.CreateDirectory(ImageFolderPath);
+
+        if (Directory.Exists(TempFolderPath))
+        {
+            // Clean out any old temp files
+            foreach (var file in Directory.GetFiles(TempFolderPath))
+            {
+                try
+                {
+                    Utilities.DebugMsg($"In InitializeFolders, deleting temp file {file}");
+                    File.Delete(file);
+                }
+                catch
+                {
+                    // nothing to do
+                }
+            }
+        }
+        else
+            Directory.CreateDirectory(TempFolderPath);
+    }
     #region Persistence Locations
     /// <summary>
     /// Indicated that a current copy is saved to app storage
@@ -210,9 +237,7 @@ public partial class Meal : ObservableObjectPlus
     // Move a suspect file into a different folder so it doesn't keep causing trouble
     public static void MoveSuspectFile(string TargetFileName)
     {
-        string suspectFolderPath = Path.Combine(MealFolderPath, SuspectFolderName);
-        Directory.CreateDirectory(suspectFolderPath);
-        File.Move(Path.Combine(MealFolderPath, TargetFileName), Path.Combine(suspectFolderPath, TargetFileName));
+        File.Move(Path.Combine(MealFolderPath, TargetFileName), Path.Combine(SuspectFolderPath, TargetFileName));
     }
     public static async Task<Meal> LoadFromRemoteAsync(MealSummary ms, bool setup = false)
     {
@@ -356,7 +381,6 @@ public partial class Meal : ObservableObjectPlus
             SavedToFile = true; //don't bother trying again until it is changed
             return;
         }
-        Directory.CreateDirectory(MealFolderPath);
         string TargetFilePath = FilePath;
         using var stream = File.Open(TargetFilePath, FileMode.Create); // Overwrites any existing file
         SaveToSnapshot();
@@ -464,5 +488,54 @@ public partial class Meal : ObservableObjectPlus
         }
         return false;
     }
-    #endregion 
+    #endregion
+    #region Archive
+    /// <summary>
+    /// Creates a ZIP archive containing the current object's data as an XML file, and optionally includes an associated
+    /// image if available.
+    /// </summary>
+    /// <remarks>The ZIP archive is saved in the application's cache directory and includes an XML file
+    /// representing the object's data. If an image is associated with the object and exists on disk, it is also
+    /// included in the archive. The method handles any exceptions internally and reports them, returning an empty
+    /// string if an error occurs.</remarks>
+    /// <returns>A string containing the full file path to the created ZIP archive. Returns an empty string if the archive could
+    /// not be created due to an error.</returns>
+    public async Task<string> ArchiveAsync()
+    {
+        Archive archive = new([this], true);
+        // Create the XML file in the cache directory
+        string xmlFileName = "DivisiBill" + archive.TimeName + ".xml";
+        string xmlFileFullname = Path.Combine(TempFolderPath, xmlFileName);
+        string zipFileFullname = Path.ChangeExtension(xmlFileFullname, ".zip");
+        try
+        {
+            using (Stream s = new FileStream(xmlFileFullname, FileMode.OpenOrCreate))
+            {
+                s.SetLength(0); // Clear the file if it exists
+                archive.AsXmlStream(s);
+                s.Flush(); // Ensure the stream is written to disk before zipping
+            }
+            using (ZipArchive archiveZip = ZipFile.Open(zipFileFullname, ZipArchiveMode.Create))
+            {
+                archiveZip.CreateEntryFromFile(xmlFileFullname, xmlFileName);
+                File.Delete(xmlFileFullname); // Delete the XML file after zipping
+                Utilities.DebugMsg($"In {nameof(ArchiveAsync)}: created zip archive {zipFileFullname} containing {xmlFileName}");
+                // Save bill image if there is one
+                if (HasImage && File.Exists(ImagePath))
+                {
+                    archiveZip.CreateEntryFromFile(ImagePath, ImageName);
+                    Utilities.DebugMsg($"In {nameof(ArchiveAsync)}: added image {ImageName} to zip archive");
+                }
+            }
+            // At this point we have a zip archive file on disk containing a single XML file containing the archive data and possibly an image file too
+            return zipFileFullname;
+        }
+        catch (Exception ex)
+        {
+            ex.ReportCrash();
+            Utilities.DebugMsg($"In {nameof(ArchiveAsync)}: exception creating zip archive {zipFileFullname}: {ex.Message}");
+            return string.Empty;
+        }
+    }
+    #endregion
 }